refactor: extract middleware functions into modules
