+++ src/index.js
@@ -3,6 +3,11 @@ import express from 'express'
 import bodyParser from 'body-parser'
 import elasticsearch from 'elasticsearch'
 
+import checkEmptyPayload from './middleware/check-empty-payload'
+import checkContentTypeIsSet from './middleware/check-content-type-is-set'
+import checkContentTypeIsJson from './middleware/check-content-type-is-json'
+import errorHandler from './middleware/error-handler'
+
 const client = new elasticsearch.Client({
   host: `${process.env.ELASTICSEARCH_PROTOCOL}://${
     process.env.ELASTICSEARCH_HOSTNAME
@@ -10,43 +15,6 @@ const client = new elasticsearch.Client({
 })
 
 const app = express()
-
-function checkEmptyPayload(req, res, next) {
-  if (
-    ['POST', 'PATCH', 'PUT'].includes(req.method) &&
-    req.headers['content-length'] === '0'
-  ) {
-    res.status(400)
-    res.set('Content-Type', 'application/json')
-    res.json({
-      message: 'Payload should not be empty'
-    })
-  }
-  next()
-}
-
-function checkContentTypeIsSet(req, res, next) {
-  if (req.headers['content-length'] !== '0' && !req.headers['content-type']) {
-    res.status(400)
-    res.set('Content-Type', 'application/json')
-    res.json({
-      message:
-        'The "Content-Type" header must be set for requests with a non-empty payload'
-    })
-  }
-  next()
-}
-
-function checkContentTypeIsJson(req, res, next) {
-  if (!req.headers['content-type'].includes('application/json')) {
-    res.status(415)
-    res.set('Content-Type', 'application/json')
-    res.json({
-      message: 'The "Content-Type" header must always be "application/json"'
-    })
-  }
-  next()
-}
 
 app.use(checkEmptyPayload)
 app.use(checkContentTypeIsSet)
@@ -99,20 +67,7 @@ app.post('/users', (req, res) => {
     })
 })
 
-app.use((err, req, res, next) => {
-  if (
-    err instanceof SyntaxError &&
-    err.status === 400 &&
-    'body' in err &&
-    err.type === 'entity.parse.failed'
-  ) {
-    res.status(400)
-    res.set('Content-Type', 'application/json')
-    res.json({ message: 'Payload should be in JSON format' })
-    return
-  }
-  next()
-})
+app.use(errorHandler)
 
 app.listen(process.env.SERVER_PORT, () => {
   // eslint-disable-next-line no-console
