+++ src/middleware/check-content-type-is-json.js
@@ -1,0 +1,12 @@
+function checkContentTypeIsJson(req, res, next) {
+  if (!req.headers['content-type'].includes('application/json')) {
+    res.status(415)
+    res.set('Content-Type', 'application/json')
+    res.json({
+      message: 'The "Content-Type" header must always be "application/json"'
+    })
+  }
+  next()
+}
+
+export default checkContentTypeIsJson
