+++ src/middleware/check-empty-payload.js
@@ -1,0 +1,15 @@
+function checkEmptyPayload(req, res, next) {
+  if (
+    ['POST', 'PATCH', 'PUT'].includes(req.method) &&
+    req.headers['content-length'] === '0'
+  ) {
+    res.status(400)
+    res.set('Content-Type', 'application/json')
+    res.json({
+      message: 'Payload should not be empty'
+    })
+  }
+  next()
+}
+
+export default checkEmptyPayload
